refactor: 소셜 로그인 SDK 전환으로 죽은 인가 코드 흐름 제거
- PKCE 조립·콜백 URL 검증·인가 URL 조립과 이를 겨누던 테스트를 함께 삭제
